pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codeapi-secagent"
version = "0.1.0"
description = "Local-first codebase+API security agent with optional Modal batching"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Security Agent", email = "secagent@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.25.0",
    "pyjwt>=2.8.0",
    "pydantic>=2.4.0",
    "typer>=0.9.0",
    "rich>=13.6.0",
    "jinja2>=3.1.0",
    "tldextract>=5.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "gitpython>=3.1.37",
    "aiofiles>=23.2.1",
    "asyncio-throttle>=1.0.2",
]

[project.optional-dependencies]
semgrep = ["semgrep>=1.45.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.9.0",
    "isort>=5.12.0",
    "mypy>=1.6.0",
]

[project.scripts]
secagent = "secagent.cli:main"

[project.urls]
Homepage = "https://github.com/example/codeapi-secagent"
Repository = "https://github.com/example/codeapi-secagent"
Issues = "https://github.com/example/codeapi-secagent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["secagent*"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
